requirements.txt
# Core dependencies for audio processing and ML models
torch>=2.0.0,<2.8.0
torchaudio>=2.0.0,<2.8.0
pyannote.audio>=3.1.0,<3.4.0
openai-whisper>=20231117
speechbrain>=0.5.16,<1.1.0
transformers>=4.30.0,<5.0.0
numpy>=1.21.0,<2.1.0
scipy>=1.7.0,<2.0.0
librosa>=0.10.0,<0.12.0
soundfile>=0.12.0,<0.14.0

# Additional required dependencies
huggingface-hub>=0.14.0
datasets>=2.0.0
torch-audiomentations>=0.11.0

seaborn>=0.11.2,<0.13.0
